# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/article/report/add" : {
      "post" : {
        "summary" : "reportArticle",
        "description" : "reportArticle",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "主键id"
                  },
                  "articleId" : {
                    "type" : "integer",
                    "description" : "被举报的文章id"
                  },
                  "reportType" : {
                    "type" : "integer",
                    "description" : "举报类型"
                  },
                  "content" : {
                    "type" : "string",
                    "description" : "举报描述"
                  },
                  "images" : {
                    "type" : "array",
                    "description" : "传入的数组图片",
                    "items" : {
                      "type" : "string"
                    }
                  },
                  "userId" : {
                    "type" : "integer",
                    "description" : "举报人id"
                  },
                  "status" : {
                    "type" : "integer",
                    "description" : "审核状态"
                  },
                  "createTime" : {
                    "type" : "string",
                    "description" : "创建时间"
                  },
                  "updateTime" : {
                    "type" : "string",
                    "description" : "更新时间"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article/report/list" : {
      "get" : {
        "summary" : "listArticleReport",
        "description" : "listArticleReport",
        "parameters" : [ {
          "name" : "pageNum",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "1"
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "10"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article/report/audit" : {
      "post" : {
        "summary" : "auditArticleReport",
        "description" : "auditArticleReport",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "status",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/article/comment/add" : {
      "post" : {
        "summary" : "addArticleComment",
        "description" : "addArticleComment",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "comment",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "parentId",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "0"
          }
        }, {
          "name" : "replyUserId",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "0"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/article/comment/audit" : {
      "post" : {
        "summary" : "auitComment",
        "description" : "auitComment",
        "parameters" : [ {
          "name" : "commentId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "reason",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "string",
            "nullable" : false,
            "default" : "false"
          }
        }, {
          "name" : "auditStatus",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/comment/report/add" : {
      "post" : {
        "summary" : "reportComment",
        "description" : "reportComment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : "主键id"
                  },
                  "commentId" : {
                    "type" : "integer",
                    "description" : "被举报的评论id"
                  },
                  "reportType" : {
                    "type" : "integer",
                    "description" : "举报类型"
                  },
                  "content" : {
                    "type" : "string",
                    "description" : "举报描述"
                  },
                  "images" : {
                    "type" : "array",
                    "description" : "传入的数组图片",
                    "items" : {
                      "type" : "string"
                    }
                  },
                  "userId" : {
                    "type" : "integer",
                    "description" : "举报人id"
                  },
                  "status" : {
                    "type" : "integer",
                    "description" : "审核状态"
                  },
                  "createTime" : {
                    "type" : "string",
                    "description" : "创建时间"
                  },
                  "updateTime" : {
                    "type" : "string",
                    "description" : "更新时间"
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comment/report/list" : {
      "get" : {
        "summary" : "listCommentReport",
        "description" : "listCommentReport",
        "parameters" : [ {
          "name" : "pageNum",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "1"
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false,
            "default" : "10"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comment/report/audit" : {
      "post" : {
        "summary" : "auditCommentReport",
        "description" : "auditCommentReport",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "status",
          "in" : "query",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/add" : {
      "post" : {
        "summary" : "addCommentLike",
        "description" : "addCommentLike",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "commentId",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "code" : {
                      "type" : "string",
                      "description" : "业务状态码  0-成功  1-失败"
                    },
                    "message" : {
                      "type" : "string",
                      "description" : "提示信息"
                    },
                    "data" : {
                      "type" : "string",
                      "description" : "响应数据"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}